server/requirements.txt
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
cloudinary==1.44.1
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.2.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
